refactor and version bump of java code
This commit pulls some of the transport code out of the Vayacondios
Itemsets client to allow for shimming up. Because the Vayacondios
client is so intimately tied to the assumption that it will work with
HTTP, it's not a clean separation at the moment. But this will at
least allow for easy shimming up for unit tests.